Various fixes. Added mission routing buttons to Mission Monitor tab.
