+++ MissionMonitor/ConfigurationWindow.xaml.cs
@@ -2,10 +2,12 @@
 using System;
 using System.Globalization;
 using System.Text.RegularExpressions;
+using System.Threading;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Data;
 using System.Windows.Input;
+using System.Windows.Media;
 
 namespace EddiMissionMonitor
 {
@@ -51,19 +53,75 @@ namespace EddiMissionMonitor
             }
         }
 
-        private void setRoute(object sender, RoutedEventArgs e)
+        private void getRoute(object sender, RoutedEventArgs e)
         {
+            Button updateButton = (Button)sender;
+            updateButton.Foreground = Brushes.Red;
+            updateButton.FontWeight = FontWeights.Bold;
+
+            Thread getRouteThread = new Thread(() =>
+            {
+                string nextSystem = missionMonitor().GetMissionsRoute();
+                Dispatcher?.Invoke(() =>
+                {
+                    updateButton.Foreground = Brushes.Black;
+                    updateButton.FontWeight = FontWeights.Regular;
+
+                    // If 'next system' found, send to clipboard
+                    if (nextSystem != null)
+                    {
+                        Clipboard.SetData(DataFormats.Text, nextSystem);
+                    }
+                });
+            })
+            {
+                IsBackground = true
+            };
+            getRouteThread.Start();
+        }
+
+        private void nextInRoute(object sender, RoutedEventArgs e)
+        {
+            string nextSystem = missionMonitor().SetNextRoute();
+
+            // If 'next system' found, send to clipboard
+            if (nextSystem != null)
+            {
+                Clipboard.SetData(DataFormats.Text, nextSystem);
+            }
 
         }
 
         private void updateRoute(object sender, RoutedEventArgs e)
         {
+            Button updateButton = (Button)sender;
+            updateButton.Foreground = Brushes.Red;
+            updateButton.FontWeight = FontWeights.Bold;
 
+            Thread updateRouteThread = new Thread(() =>
+            {
+                string nextSystem = missionMonitor().UpdateRoute();
+                Dispatcher?.Invoke(() =>
+                {
+                    updateButton.Foreground = Brushes.Black;
+                    updateButton.FontWeight = FontWeights.Regular;
+
+                    // If 'next system' found, send to clipboard
+                    if (nextSystem != null)
+                    {
+                        Clipboard.SetData(DataFormats.Text, nextSystem);
+                    }
+                });
+            })
+            {
+                IsBackground = true
+            };
+            updateRouteThread.Start();
         }
 
         private void clearRoute(object sender, RoutedEventArgs e)
         {
-
+            missionMonitor().CancelRoute();
         }
 
         private void EnsureValidInteger(object sender, TextCompositionEventArgs e)
