+++ MissionMonitor/MissionMonitor.cs
@@ -31,6 +31,8 @@ namespace EddiMissionMonitor
         private DateTime updateDat;
         public int missionsCount;
         public int? missionWarning;
+        public string missionsRouteList;
+        public decimal missionsRouteDistance;
 
         private static readonly object missionsLock = new object();
         public event EventHandler MissionUpdatedEvent;
@@ -829,7 +831,9 @@ namespace EddiMissionMonitor
             {
                 ["missions"] = new List<Mission>(missions),
                 ["missionsCount"] = missionsCount,
-                ["missionWarning"] = missionWarning
+                ["missionWarning"] = missionWarning,
+                ["routeList"] = missionsRouteList,
+                ["routeDistance"] = missionsRouteDistance
             };
             return variables;
         }
@@ -845,7 +849,10 @@ namespace EddiMissionMonitor
                     updatedat = updateDat,
                     missions = missions,
                     missionsCount = missionsCount,
-                    missionWarning = missionWarning
+                    missionWarning = missionWarning,
+                    missionsRouteList = missionsRouteList,
+                    missionsRouteDistance = missionsRouteDistance
+
                 };
                 configuration.ToFile();
             }
@@ -861,6 +868,8 @@ namespace EddiMissionMonitor
                 configuration = configuration ?? MissionMonitorConfiguration.FromFile();
                 missionsCount = configuration.missionsCount;
                 missionWarning = configuration.missionWarning ?? 60;
+                missionsRouteList = configuration.missionsRouteList;
+                missionsRouteDistance = configuration.missionsRouteDistance;
                 updateDat = configuration.updatedat;
 
                 // Build a new missions log
@@ -933,6 +942,452 @@ namespace EddiMissionMonitor
                     }
                 }
             }
+        }
+
+        public void CancelRoute()
+        {
+            // Clear route and destination variables
+            missionsRouteList = null;
+            missionsRouteDistance = 0;
+
+            EDDI.Instance.enqueueEvent(new RouteDetailsEvent(DateTime.Now, "cancel", null, missionsRouteList, 0, 0, missionsRouteDistance, null));
+        }
+
+        public string GetMissionsRoute(string homeSystem = null)
+        {
+            string nextSystem = null;
+            decimal nextDistance = 0;
+            int routeCount = 0;
+
+            List<string> systems = new List<string>();      // List of eligible mission destintaion systems
+            List<long> missionids = new List<long>();       // List of mission IDs for the next system
+
+            if (missions.Count > 0)
+            {
+                // Add current star system first
+                string currentSystem = EDDI.Instance?.CurrentStarSystem?.name;
+                systems.Add(currentSystem);
+
+                // Add origin systems for 'return to origin' missions to the 'systems' list
+                foreach (Mission mission in missions.Where(m => m.statusEDName != "Failed").ToList())
+                {
+                    if (mission.originreturn && !systems.Contains(mission.originsystem))
+                    {
+                        systems.Add(mission.originsystem);
+                    }
+                }
+
+                // Add destination systems for applicable mission types to the 'systems' list
+                foreach (Mission mission in missions.Where(m => m.statusEDName == "Active").ToList())
+                {
+                    string type = mission.typeEDName.ToLowerInvariant();
+                    switch (type)
+                    {
+                        case "assassinate":
+                        case "courier":
+                        case "delivery":
+                        case "disable":
+                        case "hack":
+                        case "massacre":
+                        case "passengerbulk":
+                        case "passengervip":
+                        case "rescue":
+                        case "salvage":
+                        case "scan":
+                        case "sightseeing":
+                        case "smuggle":
+                            {
+                                if (mission.destinationsystems == null || !mission.destinationsystems.Any())
+                                {
+                                    if (!systems.Contains(mission.destinationsystem))
+                                    {
+                                        systems.Add(mission.destinationsystem);
+                                    }
+                                }
+                                else
+                                {
+                                    foreach (DestinationSystem system in mission.destinationsystems)
+                                    {
+                                        if (!systems.Contains(system.name))
+                                        {
+                                            systems.Add(system.name);
+                                        }
+                                    }
+                                }
+                            }
+                            break;
+                    }
+                }
+
+                // Calculate the missions route using the 'Repetitive Nearest Neighbor' Algorithim (RNNA)
+                if (CalculateRNNA(systems, homeSystem))
+                {
+                    nextSystem = missionsRouteList?.Split('_')[0];
+                    nextDistance = CalculateDistance(currentSystem, nextSystem);
+                    routeCount = missionsRouteList.Split('_').Count();
+
+                    Logging.Debug("Calculated Route Selected = " + missionsRouteList + ", Total Distance = " + missionsRouteDistance);
+                    writeMissions();
+
+                    // Get mission IDs for 'next' system
+                    missionids = GetSystemMissionIds(nextSystem);
+
+                    // Set destination variables
+                    UpdateDestinationData(nextSystem, null, nextDistance);
+                }
+                else
+                {
+                    Logging.Debug("Unable to meet missions route calculation criteria");
+                }
+            }
+            EDDI.Instance.enqueueEvent(new RouteDetailsEvent(DateTime.Now, "route", nextSystem, missionsRouteList, routeCount, nextDistance, missionsRouteDistance, missionids));
+            return nextSystem;
+        }
+
+        public bool CalculateRNNA(List<string> systems, string homeSystem)
+        {
+            // Clear route list & distance
+            missionsRouteList = null;
+            missionsRouteDistance = 0;
+            bool found = false;
+
+            int numSystems = systems.Count();
+            if (numSystems > 1)
+            {
+                List<string> bestRoute = new List<string>();
+                decimal bestDistance = 0;
+
+                // Pre-load all system distances
+                if (homeSystem != null)
+                {
+                    systems.Add(homeSystem);
+                }
+                List<StarSystem> starsystems = DataProviderService.GetSystemsData(systems.ToArray(), true, false, false, false, false);
+                decimal[][] distMatrix = new decimal[systems.Count][];
+                for (int i = 0; i < systems.Count; i++)
+                {
+                    distMatrix[i] = new decimal[systems.Count];
+                }
+                for (int i = 0; i < systems.Count - 1; i++)
+                {
+                    StarSystem curr = starsystems.Find(s => s.name == systems[i]);
+                    for (int j = i + 1; j < systems.Count; j++)
+                    {
+                        StarSystem dest = starsystems.Find(s => s.name == systems[j]);
+                        decimal distance = CalculateDistance(curr, dest);
+                        distMatrix[i][j] = distance;
+                        distMatrix[j][i] = distance;
+                    }
+                }
+
+                // Repetitive Nearest Neighbor Algorithm (RNNA)
+                // Iterate through all possible routes by changing the starting system
+                for (int i = 0; i < numSystems; i++)
+                {
+                    // If starting system is a destination for a 'return to origin' mission, then not a viable route
+                    if (DestinationOriginReturn(systems[i])) { continue; }
+
+                    List<string> route = new List<string>();
+                    decimal totalDistance = 0;
+                    int currIndex = i;
+
+                    // Repeat until all systems (except starting system) are in the route
+                    while (route.Count() < numSystems - 1)
+                    {
+                        SortedList<decimal, int> nearestList = new SortedList<decimal, int>();
+
+                        // Iterate through the remaining systems to find nearest neighbor
+                        for (int j = 1; j < numSystems; j++)
+                        {
+                            // Wrap around the list
+                            int destIndex = i + j < numSystems ? i + j : i + j - numSystems;
+                            if (homeSystem != null && destIndex == 0) { destIndex = numSystems; }
+
+                            // Check if destination system previously added to the route
+                            if (route.IndexOf(systems[destIndex]) == -1)
+                            {
+                                nearestList.Add(distMatrix[currIndex][destIndex], destIndex);
+                            }
+                        }
+                        // Set the 'Nearest' system as the new 'current' system
+                        currIndex = nearestList.Values.FirstOrDefault();
+
+                        // Add 'nearest' system to the route list and add its distance to total distance traveled
+                        route.Add(systems[currIndex]);
+                        totalDistance += nearestList.Keys.FirstOrDefault();
+                    }
+
+                    // Add 'starting system' to complete the route & add its distance to total distance traveled
+                    int startIndex = homeSystem != null && i == 0 ? numSystems : i;
+                    route.Add(systems[startIndex]);
+                    if (currIndex == numSystems) { currIndex = 0; }
+                    totalDistance += distMatrix[currIndex][startIndex];
+                    Logging.Debug("Build Route Iteration #" + i + " - Route = " + string.Join("_", route) + ", Total Distance = " + totalDistance);
+
+                    // Use this route if total distance traveled is less than previous iterations
+                    if (bestDistance == 0 || totalDistance < bestDistance)
+                    {
+                        bestRoute.Clear();
+                        int homeIndex = route.IndexOf(systems[homeSystem != null ? numSystems : 0]);
+                        if (homeIndex < route.Count - 1)
+                        {
+                            // Rotate list to place homesystem at the end
+                            bestRoute = route.Skip(homeIndex + 1)
+                                .Concat(route.Take(homeIndex + 1))
+                                .ToList();
+                        }
+                        else
+                        {
+                            bestRoute = route.ToList();
+                        }
+                        bestDistance = totalDistance;
+                    }
+                }
+
+                if (bestRoute.Count == numSystems)
+                {
+                    missionsRouteList = string.Join("_", bestRoute);
+                    missionsRouteDistance = bestDistance;
+                    found = true;
+                }
+            }
+            return found;
+        }
+
+        public string SetNextRoute()
+        {
+            string nextSystem = missionsRouteList?.Split('_')[0];
+            decimal nextDistance = 0;
+            int count = 0;
+            List<long> missionids = new List<long>();       // List of mission IDs for the next system
+
+            if (nextSystem != null)
+            {
+                StarSystem curr = EDDI.Instance?.CurrentStarSystem;
+                StarSystem dest = StarSystemSqLiteRepository.Instance.GetOrFetchStarSystem(nextSystem, true);
+
+                if (dest != null && nextSystem != curr.name)
+                {
+                    nextDistance = CalculateDistance(curr, dest);
+                }
+                count = missionsRouteList.Split('_').Count();
+
+                // Get mission IDs for 'next' system
+                missionids = GetSystemMissionIds(nextSystem);
+
+                // Set destination variables
+                UpdateDestinationData(nextSystem, null, nextDistance);
+            }
+            EDDI.Instance.enqueueEvent(new RouteDetailsEvent(DateTime.Now, "next", nextSystem, missionsRouteList, count, nextDistance, missionsRouteDistance, missionids));
+            return nextSystem;
+        }
+
+        public string UpdateRoute(string updateSystem = null)
+        {
+            bool update;
+            string nextSystem = null;
+            decimal nextDistance = 0;
+            List<long> missionids = new List<long>();       // List of mission IDs for the next system
+            string currentSystem = EDDI.Instance?.CurrentStarSystem?.name;
+            List<string> route = missionsRouteList?.Split('_').ToList();
+
+            if (route.Count == 0) { update = false; }
+            else if (updateSystem == null)
+            {
+                updateSystem = route[0];
+
+                // Determine if the 'update' system in the missions route list is the current system & has no pending missions
+                update = currentSystem == updateSystem ? !SystemPendingMissions(updateSystem) : false;
+            }
+            else { update = route.Contains(updateSystem); }
+
+            // Remove 'update' system from the missions route list
+            if (update)
+            {
+                if (RemoveSystemFromRoute(updateSystem))
+                {
+                    nextSystem = missionsRouteList?.Split('_')[0];
+                    if (nextSystem != null)
+                    {
+                        nextDistance = CalculateDistance(currentSystem, nextSystem);
+
+                        // Get mission IDs for 'next' system
+                        missionids = GetSystemMissionIds(nextSystem);
+                    }
+                    Logging.Debug("Route Updated = " + missionsRouteList + ", Total Distance = " + missionsRouteDistance);
+                    writeMissions();
+
+                    // Set destination variables
+                    UpdateDestinationData(nextSystem, null, nextDistance);
+                }
+            }
+            EDDI.Instance.enqueueEvent(new RouteDetailsEvent(DateTime.Now, "update", nextSystem, missionsRouteList, route.Count, nextDistance, missionsRouteDistance, missionids));
+            return nextSystem;
+        }
+
+        private bool RemoveSystemFromRoute(string system)
+        {
+            List<string> route = missionsRouteList?.Split('_').ToList();
+            if (route.Count == 0) { return false; }
+
+            int index = route.IndexOf(system);
+            if (index > -1)
+            {
+                // Do not remove the 'home' system unless last in list
+                if (route.Count > 1 && index == route.Count - 1) { return false; }
+
+                route.RemoveAt(index);
+                if (route.Count > 0)
+                {
+                    // If other than 'next' system removed, recalculate the route
+                    if (route.Count > 2 && index > 0)
+                    {
+                        // Use copy to keep the original intact.
+                        List<string> systems = new List<string>(route);
+
+                        // Build systems list
+                        string homeSystem = systems.Last();
+                        systems.RemoveAt(systems.Count - 1);
+                        systems.Insert(0, EDDI.Instance?.CurrentStarSystem?.name);
+
+                        if (CalculateRNNA(systems, homeSystem)) { return true; }
+                    }
+                    missionsRouteList = string.Join("_", route);
+                    missionsRouteDistance = CalculateRouteDistance();
+                }
+                else
+                {
+                    missionsRouteList = null;
+                    missionsRouteDistance = 0;
+                }
+                return true;
+            }
+            return false;
+        }
+
+        public decimal CalculateDistance(string currentSystem, string destinationSystem)
+        {
+            StarSystem curr = StarSystemSqLiteRepository.Instance.GetOrFetchStarSystem(currentSystem, true);
+            StarSystem dest = StarSystemSqLiteRepository.Instance.GetOrFetchStarSystem(destinationSystem, true);
+            return CalculateDistance(curr, dest);
+        }
+
+        public decimal CalculateDistance(StarSystem curr, StarSystem dest)
+        {
+            decimal distance = -1;
+            if (curr != null && dest != null)
+            {
+                distance = (decimal)Math.Round(Math.Sqrt(Math.Pow((double)(curr.x - dest.x), 2)
+                    + Math.Pow((double)(curr.y - dest.y), 2)
+                    + Math.Pow((double)(curr.z - dest.z), 2)), 2);
+
+            }
+            return distance;
+        }
+
+        private decimal CalculateRouteDistance()
+        {
+            List<string> route = missionsRouteList?.Split('_').ToList();
+            decimal distance = 0;
+
+            if (route.Count > 0)
+            {
+                StarSystem curr = EDDI.Instance?.CurrentStarSystem;
+
+                // Get all the route coordinates from EDSM in one request
+                List<StarSystem> starsystems = DataProviderService.GetSystemsData(route.ToArray(), true, false, false, false, false);
+
+                // Get distance to the next system
+                StarSystem dest = starsystems.Find(s => s.name == route[0]);
+                distance = CalculateDistance(curr, dest);
+
+                // Calculate remaining route distance
+                for (int i = 0; i < route.Count() - 1; i++)
+                {
+                    curr = starsystems.Find(s => s.name == route[i]);
+                    dest = starsystems.Find(s => s.name == route[i + 1]);
+                    distance += CalculateDistance(curr, dest);
+                }
+            }
+            return distance;
+        }
+
+        private bool DestinationOriginReturn(string destination)
+        {
+            foreach (Mission mission in missions.Where(m => m.originreturn).ToList())
+            {
+                if (mission.destinationsystems == null)
+                {
+                    if (mission.destinationsystem == destination)
+                    {
+                        return true;
+                    }
+                }
+                else
+                {
+                    DestinationSystem system = mission.destinationsystems.FirstOrDefault(ds => ds.name == destination);
+                    if (system != null)
+                    {
+                        return true;
+                    }
+                }
+            }
+            return false;
+        }
+
+        public void SetNavigationData(string system, string station, decimal distance)
+        {
+            missionsRouteList = system;
+            missionsRouteDistance = distance;
+            UpdateDestinationData(system, station, distance);
+            writeMissions();
+        }
+
+        private bool SystemPendingMissions(string system)
+        {
+            foreach (Mission mission in missions.Where(m => m.statusEDName != "Fail").ToList())
+            {
+                string type = mission.typeEDName.ToLowerInvariant();
+                switch (type)
+                {
+                    case "assassinate":
+                    case "courier":
+                    case "delivery":
+                    case "disable":
+                    case "hack":
+                    case "massacre":
+                    case "passengerbulk":
+                    case "passengervip":
+                    case "rescue":
+                    case "salvage":
+                    case "scan":
+                    case "sightseeing":
+                    case "smuggle":
+                        {
+                            // Check if the system is origin system for 'Active' and 'Complete' missions
+                            if (mission.originsystem == system) { return true; }
+
+                            // Check if the system is destination system for 'Active' missions
+                            else if (mission.statusEDName == "Active")
+                            {
+                                if (mission.destinationsystems != null && mission.destinationsystems.Any())
+                                {
+                                    if (mission.destinationsystems.Where(d => d.name == system).Any()) { return true; }
+                                }
+                                else if (mission.destinationsystem == system) { return true; }
+                            }
+                        }
+                        break;
+                }
+            }
+            return false;
+        }
+
+        public void UpdateDestinationData(string system, string station, decimal distance)
+        {
+            EDDI.Instance.updateDestinationSystem(system);
+            EDDI.Instance.DestinationDistance = distance;
+            EDDI.Instance.updateDestinationStation(station);
         }
 
         static void RaiseOnUIThread(EventHandler handler, object sender)
